tratamento de Lista de Viagens, embarque e membros

+++ front-end/public/scripts/membros.js
@@ -5,6 +5,7 @@ import Optional from "./modules/Optional.js"
 import deleteAnInvitation from "./requests/deleteAnInvitation.js";
 import gtHeaders from "./requests/gtHeaders.js";
 import postNewMember from "./requests/postNewMember.js";
+import swal from 'sweetalert';
 
 const mmbc = membersCards();
 const urlp = urlParser();
@@ -20,6 +21,7 @@ const templateInvitedCard = document.querySelector("template#t_membro_busca");
 const urlParams = urlp.mapVariables(location.href);
 
 searchButton.addEventListener('click', function getUserAndTravelToMakeAnInvitation(){
+        inviteBlock.innerHTML = "";
 
         const urls = { "searchUser": `http://localhost:3333/usuarios/buscar?q=${searchInput.value}`,
                        "getTravel": `http://localhost:3333/viagens/ler/${urlParams.travel_id}` }
@@ -35,7 +37,7 @@ searchButton.addEventListener('click', function getUserAndTravelToMakeAnInvitati
                .then(values => ({"usuario": values[0], "viagem": values[1]}))
                .then(json => mmbc.buildToInviteCard(templateInvitedCard, json))
                .then(card => inviteBlock.appendChild(card))
-               .then(e => console.log(e))
+               .then(e => {console.log(e)})
 })
 
 window.addEventListener('load', function getActiveTravels(){
@@ -66,12 +68,39 @@ window.addEventListener('load', function getActiveTravels(){
 })
 
 window.addEventListener("guestCardCancelButtonClick", function cancelInvitation(e){
+        swal({
+                title: "Deseja cancelar este convite?",
+                text: "Você pode realizar esse convite novamente...",
+                icon: "warning",
+                buttons: true,
+                dangerMode: true,
+              })
+              .then((willDelete) => {
+                if (willDelete) {
+                        const request = deleteAnInvitation(gtHeaders.authorized(), e.detail)
 
-        const request = deleteAnInvitation(gtHeaders.authorized(), e.detail)
-
-        fetch(request.url, request.init)
-                .then(res => res.json())
-                .then(res => console.log(res.message))
+                        fetch(request.url, request.init)
+                                .then(res => res.json())
+                                .then(res => {
+                                        if(res.message !== undefined){
+                                                swal (`${res.message}` , { 
+                                                        icon: "error",
+                                                        buttons : false, 
+                                                        timer : 2000 })
+                                                .then((value) => window.location.href = `membros.html?travel_id=${urlParams.travel_id}`);
+                                        }
+                                })
+                                
+                        swal("Convite Deletado!", {
+                        icon: "success",
+                        buttons : false,
+                        timer : 2000
+                        })
+                        .then((value) => window.location.href = `membros.html?travel_id=${urlParams.travel_id}`);
+                } else {
+                        swal("Seu convite continua aguardando aprovação!")
+                }
+              })      
 })
 
 window.addEventListener("guestCardInviteButtonClick", function inviteToTravel(e){
@@ -80,5 +109,19 @@ window.addEventListener("guestCardInviteButtonClick", function inviteToTravel(e)
 
         fetch(request.url, request.init)
                 .then(res => res.json())
-                .then(res => console.log(res.message))
+                .then(res => {
+                        if(res.message !== undefined){
+                                swal (`${res.message}` , { 
+                                        icon: "error",
+                                        buttons : false, 
+                                        timer : 2000 })
+                                .then((value) => window.location.href = `membros.html?travel_id=${urlParams.travel_id}`);
+                        } else {
+                                swal ("Convite Realizado!" , { 
+                                        icon: "success",
+                                        buttons : false, 
+                                        timer : 2000 })
+                                .then((value) => window.location.href = `membros.html?travel_id=${urlParams.travel_id}`);
+                        }
+                })
 })
